// WADNR.Web/src/app/shared/generated/model/external-map-layer-detail.ts
export class ExternalMapLayerDetail { 
    ExternalMapLayerID?: number;
    DisplayName?: string | null;
    LayerUrl?: string | null;
    LayerDescription?: string | null;
    FeatureNameField?: string | null;
    DisplayOnPriorityLandscape?: boolean;
    DisplayOnProjectMap?: boolean;
    DisplayOnAllOthers?: boolean;
    IsActive?: boolean;
    IsTiledMapService?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}