// WADNR.Web/src/app/shared/generated/model/external-map-layer-upsert-request.ts
export class ExternalMapLayerUpsertRequest { 
    DisplayName?: string | null;
    LayerUrl?: string | null;
    LayerDescription?: string | null;
    FeatureNameField?: string | null;
    DisplayOnProjectMap?: boolean;
    DisplayOnPriorityLandscape?: boolean;
    DisplayOnAllOthers?: boolean;
    IsActive?: boolean;
    IsTiledMapService?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}